% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_plastics_data.R
\name{clean_plastics_data}
\alias{clean_plastics_data}
\title{Clean plastics data}
\usage{
clean_plastics_data(data = load_data(), remove_empty_country = FALSE)
}
\arguments{
\item{data}{Plastics dataset.}

\item{remove_empty_country}{Logical. If TRUE, removes rows where country is "EMPTY".}
}
\value{
A cleaned data frame.
}
\description{
Removes Grand Total rows and optionally removes empty country values.
}
