% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_country_dist.R
\name{plot_country_dist}
\alias{plot_country_dist}
\title{Plot plastic waste by country before and during COVID}
\usage{
plot_country_dist(
  data = load_data(),
  label_n = 2,
  label_largest_increase = FALSE,
  point_color = "#D95C5C",
  point_size = 3,
  point_alpha = 0.7
)
}
\arguments{
\item{data}{Plastics data.}

\item{label_n}{Number of countries to label as numeric.}

\item{label_largest_increase}{If TRUE, labels countries with the largest increase from 2019 to 2020.
If FALSE, labels countries with the largest 2020 totals.}

\item{point_color}{Color of the points.}

\item{point_size}{Size of the points as numeric.}

\item{point_alpha}{Transparency of the points as numeric.}
}
\value{
A ggplot.
}
\description{
Plot plastic waste by country before and during COVID
}
