% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_plastics_prop.R
\name{summarize_plastics_prop}
\alias{summarize_plastics_prop}
\title{Summarize changes in plastic type composition from 2019 to 2020}
\usage{
summarize_plastics_prop(data = load_data(), type = "both")
}
\arguments{
\item{data}{Plastics dataset.}

\item{type}{Type of change to return: "prop", "yoy", or "both".}
}
\value{
A tibble.
}
\description{
Summarize changes in plastic type composition from 2019 to 2020
}
