% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_country.R
\name{validate_country}
\alias{validate_country}
\title{Validate a country name}
\usage{
validate_country(country_name, data = load_data())
}
\arguments{
\item{country_name}{Country name as a character string.}

\item{data}{Plastics dataset.}
}
\value{
A validated country name.
}
\description{
Checks whether a country exists in the plastics dataset and returns the
correctly capitalized country name.
}
