% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_gt_journal.R
\name{style_gt_journal}
\alias{style_gt_journal}
\title{Style gt table with journal dashboard formatting}
\usage{
style_gt_journal(
  table,
  title,
  subtitle = NULL,
  percent = NULL,
  color_col = NULL
)
}
\arguments{
\item{table}{A gt table.}

\item{title}{Table title as string.}

\item{subtitle}{Table subtitle as string.}

\item{percent}{Column or vector of columns to format as percentages.}

\item{color_col}{Column or vector of columns to color.}
}
\value{
A styled gt table.
}
\description{
Style gt table with journal dashboard formatting
}
